Step 1
In VS Code, open the wdd230 local repository folder.
Create a new page in this root folder named "index.html".
This is your course home page. You will maintain this page throughout the course with updates and links to activities and assignments.
Create new folders (directories) named "styles", "scripts", and "images'.
These sub-folders, on the repository root directory, will contain the relevant, asset files.
Note that the words directory and folder have essentially the same meaning. Directory is the more accurate term for file systems while "folder" 📂 refers to the graphical metaphor that is generally accepted because it is highly related to the term file in the organized world.
Add a CSS file named "base.css" to the styles directory. This is your main CSS file.
Add a JS file named "getdates.js" to the scripts directory.