Step 3
Layout the page as shown in the given screenshot example.

In the base.css file, style the page with your own color schema and typography choices.
You are responsible to practice good design principles of alignment, color contrast, proximity, repetition, and consistent spacing in all of your work in this class.
Use the Google Fonts API to select one or two fonts to use on the page.
The main element has a limited width and is centered on the screen horizontally.
Layout the main column cards using CSS Grid.
Display the main sections in a single column at this point.
Set the style type of the course links list to none (do not display bullets).
Make additional CSS alterations as needed.