Step 4
In your index.html document, reference the "getdates.js" file at the end of the document, right before the closing </body> tag.
getdates.js contains the script for dynamically outputting:
the copyright year (the current year) in the footer's first paragraph, and
the date the document was last modified in the second paragraph.
Use the lastModified property of the document object to get this date/time dynamically.
You do not need to alter the document.lastModified native format.